+++ src/catalogue_api.hpp
@@ -4,12 +4,16 @@
 #include "catalogue.hpp"
 #include "http.hpp"
 
+#include <functional>
+
 namespace macha {
 
 class CatalogueApi {
     CatalogueManager& catalogue_;
+    std::function<void()> request_rescan_;
   public:
-    explicit CatalogueApi(CatalogueManager& catalogue) : catalogue_(catalogue) {}
+    explicit CatalogueApi(CatalogueManager& catalogue, std::function<void()> request_rescan = {})
+        : catalogue_(catalogue), request_rescan_(std::move(request_rescan)) {}
     HttpResponse handle(const HttpRequest&);
 };